In DataLayer.EF project refactoring
