+++ NetCoreRestApi/DataLayer.EF/IDbContext.cs
@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace DataLayer.EF
 {
-    public interface IRepositoryDbContext : IUnitOfWorkDbContext          
+    public interface IDbContext : IUnitOfWorkDbContext          
     {
         DbSet<TEntity> GetDbSet<TEntity>() where TEntity : class;
     }
