+++ NetCoreRestApi/DataLayer.EF/Repositories/BaseRepository.cs
@@ -15,7 +15,7 @@ namespace DataLayer.EF.Repositories
         protected IConverter<TEntity, TModel> Сonverter { get; private set; }
         private DbSet<TEntity> _dbSet;
 
-        public BaseRepository(IRepositoryDbContext dbContext, IConverter<TEntity, TModel> converter)
+        public BaseRepository(IDbContext dbContext, IConverter<TEntity, TModel> converter)
         {
             Сonverter = converter;
             _dbSet = dbContext.GetDbSet<TEntity>();
