+++ NetCoreRestApi/WebAPI/Startup.cs
@@ -38,9 +38,9 @@ namespace WebAPI
             services.AddTransient<IConverter<CategoryEntity, CategoryModel>, CategoryModelConverter>();
             services.AddTransient<ICategoryRepository, CategoryRepository>();
             services.AddTransient<IProductRepository, ProductRepository>();
-            services.AddTransient<IUnitOfWorkDbContext, EfDbContext>();
-            services.AddTransient<IRepositoryDbContext, EfDbContext>();
-            services.AddDbContext<EfDbContext>(opt => opt.UseInMemoryDatabase(databaseName: Guid.NewGuid().ToString()), ServiceLifetime.Singleton, ServiceLifetime.Singleton);
+            services.AddTransient<IUnitOfWorkDbContext, ProductsDbContext>();
+            services.AddTransient<IDbContext, ProductsDbContext>();
+            services.AddDbContext<ProductsDbContext>(opt => opt.UseInMemoryDatabase(databaseName: Guid.NewGuid().ToString()), ServiceLifetime.Transient, ServiceLifetime.Transient);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
